requirements.txt
bleak
pywin32
pystray
Pillow